Working f90 interface.
Added more tests to demonstrate how the f90 interface works. The matrix is
pointed to by a pointer in the C library. The pointer is translated into an
array of integers which is passed to the fortran caller. The fortran caller
can not directly access the matrix data, but has to pass the integer array to
the C library to manipulate the underlying matrix object.