Fixed for intel's ifort.
For some reason libtool when generated using FC=ifort does not know how to
compile the f90 code. I added AM_LIBTOOLFLAGS = --tag=F77 to the makefiles to
tell libtool what to do. This problem does not occur with gfortran, so maybe
there is something else not quite correct with the configure script or the
makefiles.