Cache behavior for multiply stream.
Added code for an outline of the multiply along the stream. The code keeps
track of loaded/unloaded dense blocks and counts how many blocks need to be
transferred to the GPU while multiplying. This is done assuming a GPU memory
size and evicting blocks from the GPU once they drop out of the "cache
window".