Updated memory_access.c
This version of memory_access.c implements a simple stream. It only loads the
first element of the matrix blocks. Once all three matrix streams are used,
i.e. A, B, and C, the performance drops to below 4 Gflop/s assuming that the
stream would actually calculate a matrix. The TLB misses are not the issue, as
can be seen by commenting out the separate allocation of A, B, and C, and
replacing it with a single allocation of the stream variable. In the latter
case, the 3 matrix blocks involved in the produce are next to each other in
memory and TLB should be minimally involved. However, the 2 test-cases produce
very similar TLB miss rates, indicating that it's really the total size of the
stream that affects that TLB miss rate and not so much where the blocks are.
It is important to tie the process to a core, otherwise the TLB gets flushed
whenver the process gets moved and TLB miss rates are not reproducible. An
example command line:

$ taskset -c 0 ./memory_access --TOT_CYC --TLB_DM --loops 100 -N 1000000

For larger streams I get these performance measurements:

A, B, C:  3.85 Gflop/s
A, B:     5.95 Gflop/s
A:        11.7 Gflop/s