Updated stream_multiply.c
Added Drepper's Follow type test to stream multiply. This test represents the
lower bound in terms of stream performance since it measures the time it takes
for memory access to the matrix blocks. While the computation can be
overlapped with memory access, we need at least the time it takes for memory
access to go through the stream.