I think these are VERY good results.
For this small system, you should imagine that large thresholds will be more
typically of what we hope to see for larger systems with small thresholds: by
the time we get to 1D-7, we are close to a full MM.

Its interesting that the filter approach yields a better Nel, but SpAMM does
better at Tr(P.E).  I'm not sure how to interpret that, but it seems good -- I
care a lot more about resolving the total energy than the density.

Signed-off-by: Nicolas Bock <nicolasbock@gmail.com>