Added a Frobenius norm field to SpAMM.
The matrix norm is calculated on the fly during a spamm_set() call. It is not
used so far anywhere else. The newly added test spamm_norm in tests finds a
small deviation between the norms calculated in row-major order on the dense
matrix and recursively on the SpAMM tree. This is a round-off problem.