Implemented first version of hash table quadtree.
This quadtree does not use a regular tree data structure, but a series of
hashtables that function as pointer lookup tables on each tier. While
construction of the tree is not faster than before, the direct lookup of data
on the kernel tier is O(1) with a hashtable as opposed to O(ln(N)) with a
tree.