Fixed logic in convolution.
The convolution seems to work now for the dense and the diagonal cases of the
muliply benchmark. The only known case not working right now is when the C
tree does not already have a particular C block. The convolution algorithm
should insert an empty C block in that case. This needs to be done in a way to
preserve correctness of the C tree, i.e. the tiers above the kernel tier might
need new nodes as well.