Fixed 2, i.e. moved norms out of multiply_stream_t.
  2) Can the stream element be made smaller so that the norms are not
     explicitly stored in it? The norm would have to be looked up by pointer
     reference, but that might be ok since the matrix data is looked up this
     way too.

I have changed the multiply stream element type so that it does not contain
norms anymore. Instead of referencing the block_dense data in the stream, we
are now referencing the spamm_data_t objects themselves and the norms and the
matrix data is offset directly into the spamm_data_t. This is possible because
of the align attribute in the struct definition. The attribute works with gcc,
not sure whether that's true for other compilers.