Fixed bug in k index lookup table.
In the conversion to using our own hash tables, I must have forgotten to
delete the counter increment in the loop over the matrix indices that generate
the k lookup tables. This bug manifested itself when multiplying matrices that
were 2^i+1 large. I have modifed the multiply test to use N = 513 to force the
issue.