Fixed second quicksort implementation...
in spamm_multiply.c. It takes unfortunately almost twice as long as the first
implementation, which might have to do with my somewhat hackish implementation
of getting around the fact that an "unsigned int" is, well, not signed, and
the wikipedia people assume a signed int. I am sure one could get around this
issue in a smarter way, but I have a hard time believing that it would speed
up the code by that much.