Added merge sort...
to the list sorting operations. It appears as if the quicksort is running
into the "worst case" scenario with the matrix indices. It reverts to O(N^2)
or something close to it, but definitely far away from O(N log(N)). Heap and
merge sort both have worst case O(N log(N)), and are therefore better suited,
but come with a larger prefactor than quicksort. While heap sort seems to
perform better, it is not easily parallelizable, while merge sort is.