Added an iterative version of merge sort.
Not having to recurse means that we have more external storage to keep track
of the sublist divisions, but we don't have to worry about exhausting the
stack for long lists, or call overhead for recursion. This version is slightly
faster in the synthetic list_sort benchmark than the recursive version.