The k-sort exhibits larger than N lg(N) cost.
I have added another test case to the k-sort part to verify the source of the
large cost of the k-sort. It appears as if the reason is the hashtable lookup
in the compare() to sort on the norms. We need to consider to organize the
norms better so that the hashtable lookup is not necessary in the compare().

Once we fix the overhead from the hashtable lookup in compare() we should
revivisit the issue of quicksort/mergesort. Quicksort might actually be close
to average performance, which would be N ln(N), but it is faster than merge
sort.