Found sever bug in spamm_convert().
The dense matrix was converted regardless of zero elements, i.e. the resulting
matrix tree was always full. This leads to an N^2 ln(N^2) scaling in the index
sort no matter how sparse the matrix in the test cases where we read in a
matrix from file. This commit fixes this issue.