Fixed jump table.
The jump targets are only 32-bit offsets (.long), and when loading the offset
value from the jump table (line 86), we need to make sure we only load 32-bit
and not 64. That's why we load into %eax and not %rax. In line 87 we need to
sign extend the 32-bit offset to 64-bit so we can use it in the final offset
calculation in line 89.