Orthogonal and non-orthogonal SP2 working.
For non SCF ordered test (Vancomycin) and single precision, we see times like:

48 seconds for  nonorthogonal with explicit Pao.S.Pao
7-8 seconds for nonorthogonal with (Pao.S).Pao
2 seconds for orthogonal with Por.Por

The first effect is because the explict P.S.P is completely expanded out (ie
no factoring of P.S first), and the SpAMM norm is unable to overcome the
increased complexity.  The second is because Pao fills like S^-1 and is not
less dense enough relative to S to make the extra multiply worth our while.

So, the hands down winner is Por.Por.