Changed locking in recursive multiply.
Instead of using a big lock (taskwait) in the middle of the 8 recursive
multiply calls, I added an exclusive lock around the C = C+alpha*A*B part
which only gets executed at the leaf nodes. The locks reside on the C tree and
therefore provide the minimum race condition protection. Performance on 48
cores increases quite a bit (from 35 times serial to 45 times).