Fixed bug in fortran interface.
In the interface to a C function, an array can not be declared with a deferred
shape (dimension(:, :)), because that leads the compiler to pass an array
reference and not the reference to the array data. If the interface declares
the array as a simple float, then the compiler passes the address of the first
element.