Some progress for parallel spamm_multiply().
Unfortunately the multiply segfaults on calling spamm_delete(). This is not a
parallel bug, it does this also in serial (with OpenMP disabled). I suspect
that there is a logic error in the way new nodes are allocated, which I
switched around so we can use OpenMP locks on the C matrix in order to prevent
race conditions.