Cautious and optimistic commit message:
It appears as if SP2 is actually working with SpAMM. It runs without
segfaulting, it comes out with the correct trace, and it writes something back
to disk. It remains though to check whether what is written is correct at all.
And then of course what the performance of SpAMM is compared to BCSR.