Fixed bug in stream kernel.
I didn't save some of the %r? registers on the stack and it turned out that
with optimization some of the variables in the caller (spamm_linear_multiply)
didn't get spilled to memory before calling the stream kernel. Overwriting
those registers without restoring them turns out to be pretty bad.