First very basic version of Charm++ version of SpAMM.
The multiply is not implemented yet, but there are a set() and a get() and the
chares are created hierarchically. The code lacks any kind of synchronization,
and hence doesn't actually work. At the very least there should be a barrier
of some kind between setting the matrix and printing it, and of course
eventually calculating the matrix product.