Preparing for matmul implementation.
In order to keep track of all children tasks in a matmul tier, there needs to
be a list of callbacks Node and not just a single callback. This is because in
the matmul case, a C Node instance will be involved in up to N products, where
N is the size of the matrix at this tier. The map and the lists keep track of
all children matmul tasks so that the callback method can check that this
partical branch in the convolution space is actually done.