Starting to convert program to use chare arrays.
Since single chares are not migratable, we will convert the code to use chare
arrays everywhere. The multiplication will be done using a new 3D array of
MultiplyElement chares which represent the space filling curve of the
convolution. This SFC can then be easily load balanced and distributed. This
design is much closer to what we had proposed originally in our ER.