Change Multiply constructor.
We can not have anything in the constructor that might suspend it, since the
RTS does not guarantee what method it will continue. It might not be the
constructor. If the constructor is called implicitly because another method on
the same chare was called, then this will likely cause an issue since the
method is continued instead of the constructor, which renders the chare not
fully initialized.