I might have identified the problem.
The MultiplyElement::multiply() method is suspended when it calls
Node::info(). Apparently it can happen that the load balancer migrates the
suspended chare without noticing that it is suspended. Presumably the RTS then
tries to resume the suspended chare at some point again, but since it was
migrated in the meantime and its destructor was called, it doesn't exist
anymore and the program crashes with a segfault.