This version segfaults after 4 or 5 iterations.
I have added the [threaded] attribute to Work::doSomething(), which fixed the
fixed size message version. This version uses variable size messages, and
segfaults after a few iterations (it takes some iterations for the load
balancer to kick in). The symptoms of this problem are similar to the previous
fixed size message segfault: When a thread is woken up it receives a segfault.