This version finally works!
I change the load balancing strategy from periodic to "at sync". I guess that
the problem was that the load balancer would try to migrate chares in the
middle of a reuction, and some kind of accounting inside the load balancer
database is not done correctly. Some messages are not correctly rerouted and
the RTS ends up starting threads that don't exist anymore.