Tried manual mode of load balancer.
Can't compile this code. The compiler does not see StartLB(). Maybe this is
not a very popular choice of LB? In src/ck-ldb/LBDatabase.h the two (public)
functions, LBDatabase::TurnManualLB{On,Off}() are exposed but the (private)
LBDatabase::StartLB() is not. I emailed the mailing list, I am not holding my
breath though....