First attempt at pruning code.
The code crashes using --type diagonal. The problem is that the
convolutionExists array is not passed back to the Multiply object at this
time, and calling an insert() or ckDestroy() on matrix elements that are
already there or not there is apparently problematic. The easiest way around
this would certainly be some form of introspection.