Fixed a few bugs related to arrays of CProxy.
When an array of type CProxy_* is new'ed the constructor of the proxies is not
called and the proxies are not properly NULL'ed. This can cause segfaults when
trying to later assign a ckNew'ed proxy, because the overloaded assingment
operation in the CProxy superclass will try to follow non-NULL pointers in the
left hand side of the assignment.