Clean up header files.
The config.h header file should not be installed on the system, because it
includes configure time variables that should really be included in the source
files instead. This patch removes header files from the installation such as
config.h, and includes some definitions in the sources instead of the header
files.