Moved the charmc compiler test back to the front.
Because charmc breaks a few conventions, it's a little tricky to check for it.
We can not simply set CXX to charmc and the run AC_PROG_CXX, because autoconf
will try to find the EXEEXT and fail. When called without '-o', chamrc will
not produce something like a.out, but assume that it should build only the
object file, what normally requires '-c'. We therefore now test for the normal
C++ compiler, and then switch to chamrc, unsetting a few cached values so that
autoconf will run some more checks. If we don't, then we end up with arguments
such as -MT which charmc doesn't like.