Cleaned up code again.
I think the right way to integrate SpAMM is to write a wrapper class, copying
the tree code in spammpack, including a pup() method so that the tree can be
migrated. Unfortunately, I don't think we can simply use the tree code in
spammpack. We would have to include charm++.h to get pup(), but then spammpack
would depend on charm, which we are not planning on right now.