Fixed bug in addIdentity.
For blocks in the padding region, the identity matrix that is added can not
span the full block, but should only span that part of the block that is
mapping into the actual matrix. Since blocks are agnostic as to where they are
in the matrix (for now), I added a function argument to Block::addIdentity()
to fix this.