BlockMsg is pretty broken.
We need to 'pull' the Block into the MultiplyElement, which leaves us with few
options. Parameter marshalling can't be used, since that would require us to
push the Block from Node into MultiplyElement, which would require some
serious synchronization efforts. A fixed or variable size message on the other
hand can't be used since the actual size of Block is not known until runtime.
I will try a custom packed message, which hopefully should do it.