Still broken.
Code segfaults towards the end when running

./spamm-charm --operation multiply -N 1 --verify

[node.cc:241 (blockAdd) 0.03s PE:0] tier 0 Node(0,0) creating new Block at
0x70f300
[node.cc:244 (blockAdd) 0.03s PE:0] tier 0 Node(0,0) Adding back to C with
Block at 0x70f300
tier 0 Node(0,0) Adding back to C = [
  7.059154e-01
]
[block.cc:65 (~Block) 0.03s PE:0] deleting Block at 0x7fffffffd2b0
[block.cc:65 (~Block) 0.03s PE:0] deleting Block at 0x7fffffffd280
[multiply.cc:231 (multiply) 0.03s PE:0] update norms

Program received signal SIGSEGV, Segmentation fault.
0x00007ffff700cb93 in malloc_consolidate (av=0x7ffff7211200 <main_arena>) at
memory-gnu.c:4412

The weird thing is that the Block destructor is called on a Block that was not
constructed. The Block constructore never allocates anything at
0x7fffffffd2b0. If I comment out the delete statement in the destructor the
code runs fine (albeit with a memory leak). The destructor at 0x7fffffffd2b0
is called by the blockAdd marshall, which presumably means that it is deleting
the marshalled Block. But that Block seems to be at 0x7fffffffd280.