Starting to add a new tree library.
The code will eventually be splitting a matrix into three sections. At the top
is the part that Charm++ manages and load balances. The mid and bottom
sections are stored in contiguous memory chunks so that packing is not
necessary for sending them around. The chunks are broken into a mid section,
which is managed by a recursive OpenMP enabled tree code, and a bottom section
with the linear tree code from spammpack.