Struggling with mod files.
The mod files for the threaded and the serial versions are apparently
different. At the very least they differ in the qutree:lock field, which only
exists if compiled with OpenMP. I am not sure if other differences exist. My
attempt to write serial and a threaded mod versions failed. CMake doesn't
understand where the mod files are now and the build fails...