New data structure for MatrixTree and more extensive testing
The data structure for MatrixTree has been changed. Now a Zero can
represent either a rectangular block of all zero entries or a matrix
of zero size (no entries). I have also removed Rows and Columns,
which are now represented as Rects divided into four subtrees, two
of which are Zeros of zero size. This reduces the number of data
constructors in the data type, and also reduces the coding needed
for any function on MatrixTrees.

I also added a lot more tests and put them in a "Test matrices" folder.