Changed MatrixTree accessor functions and how SpAMM checks matrices
The functions for accessing the height, width, and norm of a MatrixTree
are now called height, width, and norm instead of getHeight, getWidth,
and getNorm. This required renaming these fields in the MatrixTree
definition record syntax.

SpAMM now checks matrices for compatibility for addition or multiplication
only once at the beginning, instead of recursively at every step.

The only change in MatrixMarket is a comment.