Fix storage layout of chunk.
The components are ordered such that the matrix data comes first. The
'sequence' attribute guarantees that the storage order of the components is
the same as layed out in the type definition. All components are guaranteed to
be allocated contiguously (Section 5.3.7 F08 Standard).

The whole type is now contiguous and can potentially be aligned.