Generified 1d tree construction.
Refactored the two constructors for left and right children into one
procedure. The new procedure takes an additional argument indicating whether
it should return the left (0) or the right (1) child. The code would be even
shorter if we had arrays of pointers, but unfortunately Fortran is not so
inclined at this point.